requirements.txt
tqdm
numpy
telebot
nltk
pandas
python-dotenv

# LangChain
langchain
langchain_community
chromadb
beautifulsoup4
torch
sentence_transformers
